+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 1. [#5869](https://github.com/influxdata/chronograf/pull/5869): Optimize Alert Rules API.
 1. [#5871](https://github.com/influxdata/chronograf/pull/5871): Add TICKscipts page.
 1. [#5872](https://github.com/influxdata/chronograf/pull/5872): Open Alert Rule Builder from a TICKscript page.
+1. [#5879](https://github.com/influxdata/chronograf/pull/5879): Remove Manage Tasks page, add Alert Rules page.
 
 ### Bug Fixes